+++ checkpoint/orbax/checkpoint/_src/handlers/pytree_checkpoint_handler.py
@@ -432,7 +432,6 @@ def _get_impl_save_args(
       item=args.item,
       save_args=args.save_args,
       ocdbt_target_data_file_size=args.ocdbt_target_data_file_size,
-      enable_pinned_host_transfer=args.enable_pinned_host_transfer,
       custom_metadata=args.custom_metadata,
   )
 
@@ -482,6 +481,7 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
       array_metadata_validator: array_metadata_store_lib.Validator = (
           array_metadata_store_lib.Validator()
       ),
+      enable_pinned_host_transfer: Optional[bool] = None,
   ):
     """Creates PyTreeCheckpointHandler.
 
@@ -504,6 +504,10 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
       handler_impl: Allows overriding the internal implementation.
       pytree_metadata_options: `PyTreeMetadataOptions` to manage metadata.
       array_metadata_validator: Validator for ArrayMetadata.
+      enable_pinned_host_transfer: Whether to use pinned_host memory for the
+        transfer from device to host memory. Passing None will enable
+        pinned_host memory depending on the platform used (currently only
+        enables it for the GPU backend).
     """
     self._aggregate_handler = MsgpackHandler(
         primary_host=multiprocessing_options.primary_host,
@@ -527,6 +531,7 @@ class PyTreeCheckpointHandler(async_checkpoint_handler.AsyncCheckpointHandler):
         type_handler_registry=type_handler_registry,
         pytree_metadata_options=pytree_metadata_options,
         array_metadata_validator=array_metadata_validator,
+        enable_pinned_host_transfer=enable_pinned_host_transfer,
     )
     self._pytree_metadata_options = pytree_metadata_options
 
@@ -1058,9 +1063,6 @@ class PyTreeSaveArgs(CheckpointArgs):
       indicates no maximum file size limit.  For best results, ensure
       chunk_byte_size is smaller than this value.  For more details, refer to
       https://google.github.io/tensorstore/kvstore/ocdbt/index.html#json-kvstore/ocdbt.target_data_file_size
-    enable_pinned_host_transfer: If False, disables transfer to
-      pinned host when copying from device to host, regardless of the presence
-      of pinned host memory.
     custom_metadata: User-provided custom metadata. An arbitrary
       JSON-serializable dictionary the user can use to store additional
       information. The field is treated as opaque by Orbax.
@@ -1069,7 +1071,6 @@ class PyTreeSaveArgs(CheckpointArgs):
   item: PyTree
   save_args: Optional[PyTree] = None
   ocdbt_target_data_file_size: Optional[int] = None
-  enable_pinned_host_transfer: bool = False
   custom_metadata: tree_types.JsonType | None = None
 
   def __post_init__(self):