+++ checkpoint/CHANGELOG.md
@@ -26,6 +26,7 @@ a `PathAwaitingCreation`.
  default value.
 - #v1 Use an implementation of `load` independent of `v0.Checkpointer`.
 - #v1 Distinguish between `force` and `overwrite` save params.
+- Enable `pinned_host_transfer` for GPU backend only.
 
 ## [0.11.13] - 2025-04-30